- add app title  - refactor namings for the app  - remove generation of unique key at the rendering itself
