+++ src/Spinner/index.jsx
@@ -1,0 +1,13 @@
+import Styled from "./styles";
+
+
+function Spinner() {
+  return (
+    <Styled>
+      <div/>
+      <div/>
+    </Styled>
+  )
+}
+
+export default Spinner;
